+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 All notable changes to `@homebridge/hap-client` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## v4.0.5 (Pending Release)
+
+### Changed
+
+- fix: guard against missing username in HAP advertisement
+
+### Homebridge Dependencies
+
+- `@homebridge/hap-nodejs` @ `v2.1.3`
+
 ## v4.0.4 (2026-05-06)
 
 ### Changed